feat: Add user management with auto-generated passwords and forced password change
- Add user creation with auto-generated secure passwords
- Add admin password reset functionality for users
- Implement forced password change flow on first login
- Add must_change_password flag to user model and auth flow
- Create dedicated ChangePassword page for password change requirement
- Add migrations for must_change_password column and backups table
- Fix SQLx type mismatches (i64 vs i32 for artifact_count)
- Update docker-compose ports to high port range (30xxx)
- Configure Vite for external access via Tailscale
- Add e2e test infrastructure with global setup
- Regenerate tasks.md from speckit

+++ backend/src/api/handlers/admin.rs
@@ -58,7 +58,7 @@ pub struct BackupResponse {
     pub status: String,
     pub storage_path: Option<String>,
     pub size_bytes: i64,
-    pub artifact_count: i64,
+    pub artifact_count: i32,
     pub started_at: Option<chrono::DateTime<chrono::Utc>>,
     pub completed_at: Option<chrono::DateTime<chrono::Utc>>,
     pub error_message: Option<String>,
@@ -226,7 +226,7 @@ pub struct RestoreRequest {
 #[derive(Debug, Serialize)]
 pub struct RestoreResponse {
     pub tables_restored: Vec<String>,
-    pub artifacts_restored: i64,
+    pub artifacts_restored: i32,
     pub errors: Vec<String>,
 }
